Add cache hash note to shader ELF
When LLPC checks shader and pipeline caches, it uses the hash for
the identifier to check if required ELF files exist in the cache file.
Since the cache file keeps the mapping between the hashes and the ELF
files, we do not keep the hash value in the ELF file.

On the other hand, we need the cache creator proposed in
https://github.com/GPUOpen-Drivers/xgl/issues/64 to build the cache
for relocatable shader ELFs, but the cache creator does not have any
information about the hash. It is because the hash is generated by
LLPC (or amdllpc) and the cache creator is a seperate program.

We want to let the LLPC compiler create a new note section for the
cache hash.  It can be used for the cache creator to create the cache
file with the correct mapping between the hash and ELF.

ELF layouts before/after adding new ".note" section and its section header
for cache hash:

```
|-------------|        |-------------|
| ELF header  |        | ELF header  |
|-------------|        |-------------|
| Sections    |        | Sections    |
| ...         |        | ...         |
|-------------|  ==>   |-------------|
| Section     |        | Section     |
| headers     |        | headers     |
| ...         |        | ...         |
|-------------|        |-------------|
| Sections    |        | New section |
| ...         |---|    | header for  |
|-------------|   |    | hash note   |    This new section header's
                  |    | section    +|--| offset must be the new
   Remaining      |    |-------------|  | note section for cache hash
   sections after |    | New note    |  |
   section header |    | section for |<-/
   table must be  |    | cache hash  |
   shifted        |    |-------------|
                  \--->| Sections    |
                       | ...         |
                       |-------------|
```

The new note section for the cache hash will contain two note entries:
 1. Note name with "llpc_cache_hash" and note description with the cache hash used for the cache lookup.
 2. Note name with "llpc_version" and note description with the LLPC version - both major and minor.
    The LLPC version information will help us to understand the hash generation algorithm. We have to
    use a correct hash algorithm for the cache lookup.

For example, if the hash is "4EDBED25 ADF15238 B8C92579 423DA423" and the LLPC version is 45.4
(the major version is 45=0x2D and the minor version is 4=0x04), the new note section will be

```
.note (size = 80 bytes)
 Unknown(0)                (name = llpc_cache_hash  size = 16)
       0:4EDBED25 ADF15238 B8C92579 423DA423
 Unknown(0)                (name = llpc_version  size = 8)
       0:0000002D 00000004
```